tp: separate balance scheduler to basic and balance (#5842)
* tp: clean up stale capture table metrics


* tp: refine balance scheduler


* tp: seperate balance scheduler to basic and balance


* fix tests


* Apply suggestions from code review


* address comments
