+++ cdc/scheduler/internal/tp/scheduler_balance.go
@@ -17,46 +17,62 @@ import (
 	"math/rand"
 	"time"
 
-	"github.com/pingcap/log"
 	"github.com/pingcap/tiflow/cdc/model"
-	"go.uber.org/zap"
+	"github.com/pingcap/tiflow/pkg/config"
 )
 
-var _ scheduler = &burstBalanceScheduler{}
+var _ scheduler = &balanceScheduler{}
 
-type burstBalanceScheduler struct {
-	random *rand.Rand
+// The scheduler for balancing tables of each captures.
+type balanceScheduler struct {
+	random               *rand.Rand
+	lastRebalanceTime    time.Time
+	checkBalanceInterval time.Duration
 }
 
-func newBurstBalanceScheduler() *burstBalanceScheduler {
-	return &burstBalanceScheduler{
-		random: rand.New(rand.NewSource(time.Now().UnixNano())),
+func newBalanceScheduler(cfg *config.SchedulerConfig) *balanceScheduler {
+	return &balanceScheduler{
+		random:               rand.New(rand.NewSource(time.Now().UnixNano())),
+		checkBalanceInterval: time.Duration(cfg.CheckBalanceInterval),
 	}
 }
 
-func (b *burstBalanceScheduler) Name() string {
-	return string(schedulerTypeBurstBalance)
+func (b *balanceScheduler) Name() string {
+	return string(schedulerTypeBalance)
 }
 
-func (b *burstBalanceScheduler) Schedule(
+func (b *balanceScheduler) Schedule(
 	checkpointTs model.Ts,
 	currentTables []model.TableID,
 	captures map[model.CaptureID]*model.CaptureInfo,
 	replications map[model.TableID]*ReplicationSet,
 ) []*scheduleTask {
-	newTables := make([]model.TableID, 0)
-	intersectionTable := make(map[model.TableID]struct{})
+	now := time.Now()
+	if now.Sub(b.lastRebalanceTime) < b.checkBalanceInterval {
+		return nil
+	}
+	b.lastRebalanceTime = now
+
+	tasks := make([]*scheduleTask, 0)
+	task := buildBurstBalanceMoveTables(b.random, currentTables, captures, replications)
+	if task != nil {
+		tasks = append(tasks, task)
+	}
+	return tasks
+}
+
+func buildBurstBalanceMoveTables(
+	random *rand.Rand,
+	currentTables []model.TableID,
+	captures map[model.CaptureID]*model.CaptureInfo,
+	replications map[model.TableID]*ReplicationSet,
+) *scheduleTask {
 	captureTables := make(map[model.CaptureID][]model.TableID)
 	for _, tableID := range currentTables {
 		rep, ok := replications[tableID]
 		if !ok {
-			newTables = append(newTables, tableID)
 			continue
 		}
-		if rep.State == ReplicationSetStateAbsent {
-			newTables = append(newTables, tableID)
-		}
-		intersectionTable[tableID] = struct{}{}
 		if rep.Primary != "" {
 			captureTables[rep.Primary] = append(captureTables[rep.Primary], tableID)
 		}
@@ -64,88 +80,8 @@ func (b *burstBalanceScheduler) Schedule(
 			captureTables[rep.Secondary] = append(captureTables[rep.Secondary], tableID)
 		}
 	}
-	rmTables := make([]model.TableID, 0)
-	for tableID := range replications {
-		_, ok := intersectionTable[tableID]
-		if !ok {
-			rmTables = append(rmTables, tableID)
-		}
-	}
 
-	captureIDs := make([]model.CaptureID, 0, len(captures))
-	for captureID := range captures {
-		captureIDs = append(captureIDs, captureID)
-	}
-	// TODO support table re-balance when adding a new capture.
-	tasks := make([]*scheduleTask, 0)
-	if len(newTables) != 0 {
-		tasks = append(
-			tasks, newBurstBalanceAddTables(checkpointTs, newTables, captureIDs))
-		if len(newTables) == len(currentTables) {
-			return tasks
-		}
-	}
-	if len(rmTables) > 0 {
-		tasks = append(
-			tasks, newBurstBalanceRemoveTables(rmTables, replications))
-	}
-
-	// Skip rebalance if there are some table added or removed.
-	if len(tasks) == 0 {
-		// We does not need accept callback here.
-		accept := (callback)(nil)
-		task := newBurstBalanceMoveTables(accept, b.random, currentTables, captures, replications)
-		if task != nil {
-			tasks = append(tasks, task)
-		}
-	}
-
-	return tasks
-}
-
-func newBurstBalanceAddTables(
-	checkpointTs model.Ts, newTables []model.TableID, captureIDs []model.CaptureID,
-) *scheduleTask {
-	idx := 0
-	tables := make([]addTable, 0, len(newTables))
-	for _, tableID := range newTables {
-		tables = append(tables, addTable{
-			TableID:      tableID,
-			CaptureID:    captureIDs[idx],
-			CheckpointTs: checkpointTs,
-		})
-		idx++
-		if idx >= len(captureIDs) {
-			idx = 0
-		}
-	}
-	return &scheduleTask{burstBalance: &burstBalance{
-		AddTables: tables,
-	}}
-}
-
-func newBurstBalanceRemoveTables(
-	rmTables []model.TableID, replications map[model.TableID]*ReplicationSet,
-) *scheduleTask {
-	tables := make([]removeTable, 0, len(rmTables))
-	for _, tableID := range rmTables {
-		rep := replications[tableID]
-		var captureID model.CaptureID
-		if rep.Primary != "" {
-			captureID = rep.Primary
-		} else if rep.Secondary != "" {
-			captureID = rep.Secondary
-		} else {
-			log.Warn("tpscheduler: primary or secondary not found for removed table",
-				zap.Any("table", rep))
-			continue
-		}
-		tables = append(tables, removeTable{
-			TableID:   tableID,
-			CaptureID: captureID,
-		})
-	}
-	return &scheduleTask{burstBalance: &burstBalance{
-		RemoveTables: tables,
-	}}
+	// We does not need accept callback here.
+	accept := (callback)(nil)
+	return newBurstBalanceMoveTables(accept, random, captures, replications)
 }
